[webgpu] And int64 to cast
This PR adds the int64 type to cast to avoid data readback to cpu for
some models during the execution.

It won't bring the perf regression since no matter the cast is from
int64 to int32 or int32 to int64, usually the other type is webgpu supported which
means that the previous op or the following op runs on gpu.